feat(trusty-mpm): opt-in tmux.alternate_screen config knob
The alternate screen buffer, not history-limit, is why tm sessions have no
scrollback. tm sets history-limit to 100,000 and verifies the readback, and
panes hold thousands of lines — but a full-screen TUI draws into the alternate
screen, which tmux discards rather than appending to the pane's history.
Measured: panes with the Claude binary rendering report alternate_on=1; plain
shells report alternate_on=0. tm never set alternate-screen at all.

Adds `tmux.alternate_screen` to ~/.trusty-tools/trusty-mpm/config.yaml
alongside history_limit and mouse, defaulting to `on` — tmux's factory value
and today's behaviour. A configured `off` is what turns it off.

alternate-screen is a pane option inheriting from the window scope, not a
server option. Against live tmux 3.6b, `set-option -pg alternate-screen off`
and `set-option -s alternate-screen off` both exit 0, a same-flag readback
reports "off", and the pane still enters the alternate screen; only `-wg`
takes effect. The new SetWindowGlobalOption / ShowWindowGlobalOption pair
keeps the set and its verification probe on the same scope, so a wrong-scope
set cannot verify green. The option rides the existing #3386 apply-and-verify
path before new-session.

The cost of turning it off is documented on TmuxConfig::alternate_screen: the
tmux server is shared, so vim, less, htop and man also stop restoring the
screen they covered, each leaving its final frame in the scrollback. It is
not retroactive.

Inline `mod tests` blocks in the three touched files are split into sibling
_tests.rs files — the basename made them count as production SLOC and all
three crossed the 500 cap.

Closes #5151
