+++ crates/trusty-agents/src/tmux/orchestrator.rs
@@ -137,6 +137,7 @@ impl TmuxOrchestrator {
             dir,
             trusty_common::tmux::DEFAULT_TMUX_HISTORY_LIMIT,
             trusty_common::tmux::DEFAULT_TMUX_MOUSE,
+            trusty_common::tmux::DEFAULT_TMUX_ALTERNATE_SCREEN,
             false,
             None,
         )
