+++ crates/trusty-common/changelog.d/5151-alternate-screen-window-scope.md
@@ -1,0 +1,4 @@
+Added
+
+- `tmux::scrollback_option_commands` and `tmux::managed_session_commands` now emit an `alternate-screen` entry alongside `history-limit` and `mouse`, and both take an `alternate_screen: bool` parameter ([#5151](https://github.com/bobmatnyc/trusty-tools/issues/5151)). `DEFAULT_TMUX_ALTERNATE_SCREEN` is `true` — tmux's own factory value — so nothing changes until a consumer passes `false`.
+- Two `TmuxCommand` variants for the window option scope: `SetWindowGlobalOption` (`set-option -wg`) and `ShowWindowGlobalOption` (`show-options -wg -v`). `alternate-screen` is a pane option inheriting from the window scope, not a server option like `history-limit`. Measured against tmux 3.6b, `set-option -pg alternate-screen off` and `set-option -s alternate-screen off` both exit 0, a same-flag readback reports `off`, and the pane still enters the alternate screen; only the window scope takes effect. Pairing the two variants keeps a caller's set and its verification probe on the same scope.
