+++ crates/trusty-mpm/src/core/tmux.rs
@@ -65,9 +65,10 @@
 use tracing::warn;
 
 pub use trusty_common::tmux::{
-    DEFAULT_TMUX_HISTORY_LIMIT, DEFAULT_TMUX_MOUSE, HISTORY_LIMIT_OPTION, MOUSE_OPTION,
-    PANE_LIST_FORMAT, SESSION_LIST_FORMAT, TmuxCommand, TmuxTarget, WINDOW_LIST_FORMAT,
-    managed_session_commands, scrollback_option_commands, tmux_argv,
+    ALTERNATE_SCREEN_OPTION, DEFAULT_TMUX_ALTERNATE_SCREEN, DEFAULT_TMUX_HISTORY_LIMIT,
+    DEFAULT_TMUX_MOUSE, HISTORY_LIMIT_OPTION, MOUSE_OPTION, PANE_LIST_FORMAT, SESSION_LIST_FORMAT,
+    TmuxCommand, TmuxTarget, WINDOW_LIST_FORMAT, managed_session_commands,
+    scrollback_option_commands, tmux_argv,
 };
 
 /// Resolve the `tmux` binary, preferring live `PATH` and falling back to
@@ -175,8 +176,17 @@ fn managed_session_command_sequence(
     workdir: Option<&str>,
     history_limit: u32,
     mouse: bool,
+    alternate_screen: bool,
 ) -> Vec<TmuxCommand> {
-    managed_session_commands(name, workdir, history_limit, mouse, true, None)
+    managed_session_commands(
+        name,
+        workdir,
+        history_limit,
+        mouse,
+        alternate_screen,
+        true,
+        None,
+    )
 }
 
 /// Outcome of [`create_managed_session`] (#3386 review finding).
@@ -282,11 +292,22 @@ pub fn create_managed_session(
 
     let config = crate::core::trusty_tools_config::TrustyToolsConfig::load();
     let opts = crate::core::trusty_tools_config::resolve_tmux_options(&config);
-    let commands = managed_session_command_sequence(name, workdir, opts.history_limit, opts.mouse);
+    let commands = managed_session_command_sequence(
+        name,
+        workdir,
+        opts.history_limit,
+        opts.mouse,
+        opts.alternate_screen,
+    );
     let split = commands.len().saturating_sub(1);
     let (options, new_session) = commands.split_at(split);
 
-    let options_verified = apply_and_verify_scrollback_options(bin, options, opts.history_limit);
+    let options_verified = apply_and_verify_scrollback_options(
+        bin,
+        options,
+        opts.history_limit,
+        opts.alternate_screen,
+    );
     if !options_verified {
         warn!(
             "#3386: tmux scrollback ergonomics could not be verified after \
@@ -362,7 +383,8 @@ pub(crate) fn ensure_server_up(bin: &str) -> Result<(), String> {
 }
 
 /// Run [`ensure_server_up`], apply every entry of `options`, then confirm
-/// via [`probe_history_limit`] — retrying the WHOLE cycle up to
+/// via [`probe_history_limit`] and [`probe_alternate_screen`] — retrying the
+/// WHOLE cycle up to
 /// [`APPLY_VERIFY_MAX_ATTEMPTS`] times (#3386 review finding): a successful
 /// `start-server` does NOT by itself prove the FOLLOWING `set-option -g`
 /// landed (the server could still be torn down between the two calls, or
@@ -374,21 +396,31 @@ pub(crate) fn ensure_server_up(bin: &str) -> Result<(), String> {
 /// `new-session` believing the ergonomics landed when they did not — the
 /// original #3386 bug was exactly a `warn!`-and-continue on a single failed
 /// attempt.
-/// What: returns `true` as soon as one cycle's probe confirms
-/// `history-limit == expected_history_limit`; returns `false` once
-/// [`APPLY_VERIFY_MAX_ATTEMPTS`] cycles are exhausted without a confirmed
-/// match. Every failure within a cycle is logged via `warn!` naming #3386;
-/// never panics, never blocks the caller from proceeding to `new-session`
-/// either way — the return value is the caller's signal to surface the
-/// degraded outcome, not a reason to abort session creation.
+/// What: returns `true` as soon as one cycle's probes confirm BOTH
+/// `history-limit == expected_history_limit` (global scope) and
+/// `alternate-screen == expected_alternate_screen` (window scope, #5151);
+/// returns `false` once [`APPLY_VERIFY_MAX_ATTEMPTS`] cycles are exhausted
+/// without a confirmed match on both. Every failure within a cycle is logged
+/// via `warn!`; never panics, never blocks the caller from proceeding to
+/// `new-session` either way — the return value is the caller's signal to
+/// surface the degraded outcome, not a reason to abort session creation.
+///
+/// Each probe reads the SAME tmux option scope its `set-option` wrote —
+/// `-g` for `history-limit`, `-wg` for `alternate-screen`. A probe that reads
+/// a different scope than it wrote would report success for a `set-option`
+/// that silently did nothing (measured: `set-option -pg alternate-screen off`
+/// exits 0 and leaves the pane still on the alternate screen), which is the
+/// same fail-open shape #3386 was filed against.
 /// Test: `apply_and_verify_scrollback_options_succeeds_on_second_attempt`,
-/// `apply_and_verify_scrollback_options_returns_false_after_exhausting_retries`
-/// (both drive a scripted fake `tmux` binary — no live tmux server
-/// required).
+/// `apply_and_verify_scrollback_options_returns_false_after_exhausting_retries`,
+/// `apply_and_verify_scrollback_options_fails_when_only_alternate_screen_is_wrong`,
+/// `apply_and_verify_probes_the_same_scope_it_sets` (all drive a scripted
+/// fake `tmux` binary — no live tmux server required).
 fn apply_and_verify_scrollback_options(
     bin: &str,
     options: &[TmuxCommand],
     expected_history_limit: u32,
+    expected_alternate_screen: bool,
 ) -> bool {
     for attempt in 1..=APPLY_VERIFY_MAX_ATTEMPTS {
         if let Err(e) = ensure_server_up(bin) {
@@ -415,17 +447,45 @@ fn apply_and_verify_scrollback_options(
                     Ok(_) => {}
                 }
             }
-            match probe_history_limit(bin) {
-                Ok(observed) if observed == expected_history_limit => return true,
-                Ok(observed) => warn!(
-                    "#3386: tmux history-limit verification mismatch on apply-and-verify \
-                     attempt {attempt}/{APPLY_VERIFY_MAX_ATTEMPTS} — expected \
-                     {expected_history_limit}, observed {observed}"
-                ),
-                Err(e) => warn!(
-                    "#3386: tmux history-limit verification probe failed on apply-and-verify \
-                     attempt {attempt}/{APPLY_VERIFY_MAX_ATTEMPTS}: {e}"
-                ),
+            let history_ok = match probe_history_limit(bin) {
+                Ok(observed) if observed == expected_history_limit => true,
+                Ok(observed) => {
+                    warn!(
+                        "#3386: tmux history-limit verification mismatch on apply-and-verify \
+                         attempt {attempt}/{APPLY_VERIFY_MAX_ATTEMPTS} — expected \
+                         {expected_history_limit}, observed {observed}"
+                    );
+                    false
+                }
+                Err(e) => {
+                    warn!(
+                        "#3386: tmux history-limit verification probe failed on apply-and-verify \
+                         attempt {attempt}/{APPLY_VERIFY_MAX_ATTEMPTS}: {e}"
+                    );
+                    false
+                }
+            };
+            // #5151: read back the WINDOW scope the alternate-screen set wrote.
+            let alternate_screen_ok = match probe_alternate_screen(bin) {
+                Ok(observed) if observed == expected_alternate_screen => true,
+                Ok(observed) => {
+                    warn!(
+                        "#5151: tmux alternate-screen verification mismatch on apply-and-verify \
+                         attempt {attempt}/{APPLY_VERIFY_MAX_ATTEMPTS} — expected \
+                         {expected_alternate_screen}, observed {observed}"
+                    );
+                    false
+                }
+                Err(e) => {
+                    warn!(
+                        "#5151: tmux alternate-screen verification probe failed on \
+                         apply-and-verify attempt {attempt}/{APPLY_VERIFY_MAX_ATTEMPTS}: {e}"
+                    );
+                    false
+                }
+            };
+            if history_ok && alternate_screen_ok {
+                return true;
             }
         }
         if attempt < APPLY_VERIFY_MAX_ATTEMPTS {
@@ -459,6 +519,46 @@ fn probe_history_limit(bin: &str) -> Result<u32, String> {
             let raw = String::from_utf8_lossy(&output.stdout).trim().to_string();
             raw.parse::<u32>()
                 .map_err(|_| format!("unparsable show-options output: {raw:?}"))
+        }
+        Ok(output) => Err(String::from_utf8_lossy(&output.stderr).into_owned()),
+        Err(e) => Err(e.to_string()),
+    }
+}
+
+/// Read back the tmux server's current WINDOW-scoped `alternate-screen`
+/// option (#5151).
+///
+/// Why: the verification counterpart to the
+/// `set-option -wg alternate-screen` entry
+/// [`scrollback_option_commands`](trusty_common::tmux::scrollback_option_commands)
+/// emits, and deliberately scope-matched to it. `alternate-screen` is a
+/// window/pane option, not a server option like `history-limit`; probing
+/// `show-options -g` (or `-pg`) here would read a scope the `set` never
+/// wrote, which is how a wrong-scope `set-option` verifies green while the
+/// pane still enters the alternate screen.
+/// What: runs `tmux show-options -wg -v alternate-screen` and maps the single
+/// line of stdout — `on` → `true`, `off` → `false`. Any spawn failure,
+/// non-zero exit, or unrecognised value is returned as `Err`; a value tmux
+/// does not document (anything but `on`/`off`) is NOT guessed at, because a
+/// wrong guess here is a silent pass.
+/// Test: `probe_alternate_screen_reads_back_on`,
+/// `probe_alternate_screen_reads_back_off`,
+/// `probe_alternate_screen_errors_on_unrecognised_value`,
+/// `probe_alternate_screen_errors_on_nonzero_exit`.
+fn probe_alternate_screen(bin: &str) -> Result<bool, String> {
+    match run_tmux_with_bin(
+        bin,
+        &TmuxCommand::ShowWindowGlobalOption {
+            name: ALTERNATE_SCREEN_OPTION.to_string(),
+        },
+    ) {
+        Ok(output) if output.status.success() => {
+            let raw = String::from_utf8_lossy(&output.stdout).trim().to_string();
+            match raw.as_str() {
+                "on" => Ok(true),
+                "off" => Ok(false),
+                _ => Err(format!("unrecognised show-options output: {raw:?}")),
+            }
         }
         Ok(output) => Err(String::from_utf8_lossy(&output.stderr).into_owned()),
         Err(e) => Err(e.to_string()),
@@ -522,298 +622,5 @@ pub fn send_line(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn resolve_tmux_binary_does_not_panic() {
-        // Works whether or not tmux is installed on the test host.
-        let _ = resolve_tmux_binary();
-    }
-
-    #[test]
-    fn resolve_tmux_binary_or_bare_never_empty() {
-        // Even when resolution fails entirely, the bare "tmux" fallback keeps
-        // this non-empty — callers always get a spawnable binary name.
-        assert!(!resolve_tmux_binary_or_bare().is_empty());
-    }
-
-    #[test]
-    fn run_tmux_with_bin_does_not_panic_on_missing_binary() {
-        // A deliberately bogus binary name must surface as an Err, never panic.
-        let result = run_tmux_with_bin(
-            "definitely-not-a-real-tmux-binary-2398",
-            &TmuxCommand::ListSessions,
-        );
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn create_managed_session_does_not_panic_on_missing_binary() {
-        // Same smoke guarantee for the session-creation choke point: a bogus
-        // binary must degrade to an Err (from the final new-session spawn),
-        // never panic — including through the best-effort scrollback loop.
-        let result = create_managed_session(
-            Some("definitely-not-a-real-tmux-binary-2398"),
-            "tmpm-test-2398-no-bin",
-            None,
-        );
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn managed_session_command_sequence_matches_shared_layer() {
-        // #3004 integration assertion: this crate's config-resolved
-        // sequence must be EXACTLY what `trusty_common::tmux` would build
-        // for `idempotent: true, command: None` (trusty-mpm's creation
-        // semantics) with the same parameters — proving trusty-mpm's
-        // session creation genuinely routes through the shared layer
-        // rather than a local re-implementation drifting from it.
-        let got = managed_session_command_sequence("tmpm-sess", Some("/tmp/proj"), 100_000, true);
-        let want = trusty_common::tmux::managed_session_commands(
-            "tmpm-sess",
-            Some("/tmp/proj"),
-            100_000,
-            true,
-            true,
-            None,
-        );
-        assert_eq!(got, want);
-    }
-
-    #[test]
-    fn managed_session_command_sequence_applies_options_before_new_session() {
-        let cmds = managed_session_command_sequence("tmpm-sess", None, 50_000, false);
-        assert_eq!(cmds.len(), 3);
-        assert!(matches!(cmds[0], TmuxCommand::SetGlobalOption { .. }));
-        assert!(matches!(cmds[1], TmuxCommand::SetGlobalOption { .. }));
-        assert!(matches!(cmds[2], TmuxCommand::NewSession { .. }));
-    }
-
-    // ── #3386: server-up confirmed before any `set-option -g` ───────────
-    //
-    // All of these drive a scripted fake `tmux` binary (a `#!/bin/sh` script
-    // written to a temp dir) rather than a live tmux server, so they are
-    // deterministic and safe to run anywhere `sh` exists.
-
-    /// Write an executable shell script at `dir/name` with body `body`,
-    /// returning its path as a `String` suitable for `create_managed_session`
-    /// / `run_tmux_with_bin`'s `bin` parameter.
-    fn write_fake_tmux(dir: &std::path::Path, name: &str, body: &str) -> String {
-        use std::os::unix::fs::PermissionsExt;
-        let path = dir.join(name);
-        std::fs::write(&path, body).expect("write fake tmux script");
-        let mut perms = std::fs::metadata(&path)
-            .expect("stat fake tmux script")
-            .permissions();
-        perms.set_mode(0o755);
-        std::fs::set_permissions(&path, perms).expect("chmod fake tmux script");
-        path.to_string_lossy().into_owned()
-    }
-
-    #[test]
-    fn ensure_server_up_retries_then_succeeds() {
-        // Fails the first two `start-server` invocations (simulating the
-        // #3386 window right after a dead server's socket dir is torn down),
-        // then succeeds on the third — exactly `START_SERVER_MAX_ATTEMPTS`.
-        let dir = tempfile::tempdir().unwrap();
-        let counter = dir.path().join("attempts");
-        std::fs::write(&counter, "0").unwrap();
-        let script = format!(
-            "#!/bin/sh\nn=$(cat '{counter}')\nn=$((n + 1))\necho \"$n\" > '{counter}'\nif [ \"$n\" -lt 3 ]; then\n  echo 'error connecting to socket (No such file or directory)' >&2\n  exit 1\nfi\nexit 0\n",
-            counter = counter.display()
-        );
-        let bin = write_fake_tmux(dir.path(), "fake-tmux-retry", &script);
-
-        let result = ensure_server_up(&bin);
-
-        assert!(
-            result.is_ok(),
-            "must eventually succeed within {START_SERVER_MAX_ATTEMPTS} attempts: {result:?}"
-        );
-        assert_eq!(
-            std::fs::read_to_string(&counter).unwrap().trim(),
-            "3",
-            "must have retried exactly up to the succeeding attempt"
-        );
-    }
-
-    #[test]
-    fn ensure_server_up_fails_loudly_after_exhausting_retries() {
-        // Always fails — the probe must return a loud Err after exhausting
-        // every attempt rather than silently reporting success.
-        let dir = tempfile::tempdir().unwrap();
-        let script = "#!/bin/sh\necho 'error connecting to socket (No such file or directory)' >&2\nexit 1\n";
-        let bin = write_fake_tmux(dir.path(), "fake-tmux-always-fails", script);
-
-        let result = ensure_server_up(&bin);
-
-        assert!(
-            result.is_err(),
-            "an always-failing server probe must error loudly, never proceed as if it succeeded"
-        );
-        assert!(
-            result.unwrap_err().contains("error connecting to socket"),
-            "the Err must carry the underlying failure reason"
-        );
-    }
-
-    #[test]
-    fn probe_history_limit_reads_back_configured_value() {
-        let dir = tempfile::tempdir().unwrap();
-        let script = "#!/bin/sh\necho 100000\nexit 0\n";
-        let bin = write_fake_tmux(dir.path(), "fake-tmux-show-options", script);
-
-        assert_eq!(probe_history_limit(&bin), Ok(100_000));
-    }
-
-    #[test]
-    fn probe_history_limit_errors_on_unparsable_output() {
-        let dir = tempfile::tempdir().unwrap();
-        let script = "#!/bin/sh\necho 'not-a-number'\nexit 0\n";
-        let bin = write_fake_tmux(dir.path(), "fake-tmux-bad-output", script);
-
-        let result = probe_history_limit(&bin);
-        assert!(
-            result.is_err(),
-            "unparsable show-options output must error loudly, not silently pass verification"
-        );
-    }
-
-    #[test]
-    fn probe_history_limit_errors_on_nonzero_exit() {
-        let dir = tempfile::tempdir().unwrap();
-        let script = "#!/bin/sh\necho 'no server running' >&2\nexit 1\n";
-        let bin = write_fake_tmux(dir.path(), "fake-tmux-show-options-fails", script);
-
-        assert!(probe_history_limit(&bin).is_err());
-    }
-
-    #[test]
-    fn create_managed_session_confirms_server_before_applying_options() {
-        // #3386 end-to-end (still no live tmux): a fake `tmux` that records
-        // each invocation's tmux sub-command (argv[0]) to a log file, always
-        // succeeds, and echoes back the CONFIGURED history-limit for
-        // `show-options` so the apply-and-verify cycle matches on its first
-        // attempt. This exercises the EXACT choke point the
-        // resume-after-server-death path (`SessionManager::resume` →
-        // `resume_workdir::create_and_verify_pane` → `RealTmuxDriver::
-        // create_session` → `TmuxDriver::create_session`) routes through, so
-        // asserting the recorded order here proves the fix for that path too.
-        //
-        // The expected history-limit is read via the SAME config-resolution
-        // path `create_managed_session` itself uses (rather than a hardcoded
-        // constant) so this test stays correct regardless of the host's own
-        // `~/.trusty-mpm` config.
-        let dir = tempfile::tempdir().unwrap();
-        let log = dir.path().join("calls.log");
-        let config = crate::core::trusty_tools_config::TrustyToolsConfig::load();
-        let opts = crate::core::trusty_tools_config::resolve_tmux_options(&config);
-        let script = format!(
-            "#!/bin/sh\necho \"$1\" >> '{log}'\ncase \"$1\" in\n  show-options) echo {history_limit} ;;\nesac\nexit 0\n",
-            log = log.display(),
-            history_limit = opts.history_limit
-        );
-        let bin = write_fake_tmux(dir.path(), "fake-tmux-order", &script);
-
-        let outcome = create_managed_session(Some(&bin), "tmpm-3386-order-test", None)
-            .expect("fake tmux always exits 0");
-
-        assert!(
-            outcome.options_verified,
-            "a fake tmux that echoes back the configured history-limit must verify successfully"
-        );
-
-        let calls = std::fs::read_to_string(&log).unwrap();
-        let lines: Vec<&str> = calls.lines().collect();
-
-        assert_eq!(
-            lines.first(),
-            Some(&"start-server"),
-            "the server must be confirmed up BEFORE any set-option -g (#3386): {lines:?}"
-        );
-        let new_session_pos = lines
-            .iter()
-            .position(|l| *l == "new-session")
-            .expect("new-session must still run");
-        let set_option_positions: Vec<usize> = lines
-            .iter()
-            .enumerate()
-            .filter(|(_, l)| **l == "set-option")
-            .map(|(i, _)| i)
-            .collect();
-        assert!(
-            !set_option_positions.is_empty(),
-            "the scrollback/mouse options must still be applied: {lines:?}"
-        );
-        assert!(
-            set_option_positions
-                .iter()
-                .all(|&p| p > 0 && p < new_session_pos),
-            "set-option -g must run strictly AFTER start-server and BEFORE new-session \
-             (#3386 / pre-existing #2398 ordering guarantee): {lines:?}"
-        );
-    }
-
-    // ── #3386 review: apply-and-verify retries the WHOLE cycle ──────────
-
-    #[test]
-    fn apply_and_verify_scrollback_options_succeeds_on_second_attempt() {
-        // The probe reports the WRONG value on the first cycle, then the
-        // correct value from the second cycle onward — the whole
-        // apply-and-verify cycle (not just start-server) must retry and
-        // eventually confirm success.
-        let dir = tempfile::tempdir().unwrap();
-        let counter = dir.path().join("probe-attempts");
-        std::fs::write(&counter, "0").unwrap();
-        let script = format!(
-            "#!/bin/sh\ncase \"$1\" in\n  show-options)\n    n=$(cat '{counter}')\n    n=$((n + 1))\n    echo \"$n\" > '{counter}'\n    if [ \"$n\" -lt 2 ]; then\n      echo 2000\n    else\n      echo 100000\n    fi\n    ;;\nesac\nexit 0\n",
-            counter = counter.display()
-        );
-        let bin = write_fake_tmux(dir.path(), "fake-tmux-verify-retry", &script);
-        let options = trusty_common::tmux::scrollback_option_commands(100_000, true);
-
-        let verified = apply_and_verify_scrollback_options(&bin, &options, 100_000);
-
-        assert!(
-            verified,
-            "must succeed once the probe reports the expected value on a later attempt"
-        );
-        assert_eq!(
-            std::fs::read_to_string(&counter).unwrap().trim(),
-            "2",
-            "must have retried the WHOLE cycle (not just start-server) exactly once"
-        );
-    }
-
-    #[test]
-    fn apply_and_verify_scrollback_options_returns_false_after_exhausting_retries() {
-        // The probe ALWAYS reports the wrong value — every apply-and-verify
-        // cycle must be attempted (never silently assumed to have worked),
-        // and the function must return `false` (the caller-visible degraded
-        // signal) rather than `true` once every attempt is exhausted.
-        let dir = tempfile::tempdir().unwrap();
-        let counter = dir.path().join("probe-attempts");
-        std::fs::write(&counter, "0").unwrap();
-        let script = format!(
-            "#!/bin/sh\ncase \"$1\" in\n  show-options)\n    n=$(cat '{counter}')\n    n=$((n + 1))\n    echo \"$n\" > '{counter}'\n    echo 2000\n    ;;\nesac\nexit 0\n",
-            counter = counter.display()
-        );
-        let bin = write_fake_tmux(dir.path(), "fake-tmux-verify-always-wrong", &script);
-        let options = trusty_common::tmux::scrollback_option_commands(100_000, true);
-
-        let verified = apply_and_verify_scrollback_options(&bin, &options, 100_000);
-
-        assert!(
-            !verified,
-            "must return false (never silently proceed as verified) once every \
-             apply-and-verify attempt is exhausted — this is the #3386 review's caller-visible \
-             degraded signal"
-        );
-        assert_eq!(
-            std::fs::read_to_string(&counter).unwrap().trim(),
-            APPLY_VERIFY_MAX_ATTEMPTS.to_string(),
-            "must have made exactly APPLY_VERIFY_MAX_ATTEMPTS probe attempts"
-        );
-    }
-}
+#[path = "tmux_tests.rs"]
+mod tmux_tests;
